rename viewmodel
