+++ Src/MoneyFox.Ui/Views/Categories/CategorySelection/DesktopSelectedCategoryPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 
 <ContentPage
-    x:Class="MoneyFox.Ui.Views.Categories.DesktopSelectedCategoryPage"
+    x:Class="MoneyFox.Ui.Views.Categories.CategorySelection.DesktopSelectedCategoryPage"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:resources="clr-namespace:MoneyFox.Ui.Resources.Strings"
